Move _maybe_indices_to_slice to multiindex.py.
